fix(condo): DOMA-8414 removed hook to useCompletedTourModals and returns modals from it
