+++ apps/condo/domains/onboarding/hooks/TourContext/useCompletedTourModals.tsx
@@ -1,11 +1,19 @@
 import { TourStepStatusType, TourStepTypeType } from '@app/condo/schema'
+import styled from '@emotion/styled'
+import { Col, Row } from 'antd'
 import get from 'lodash/get'
+import isEmpty from 'lodash/isEmpty'
+import pickBy from 'lodash/pickBy'
+import Link from 'next/link'
 import { useRouter } from 'next/router'
 import React, { useCallback, useMemo, useState } from 'react'
 
-import { Building, Guide, LayoutList, Meters, Wallet } from '@open-condo/icons'
+import { Building, Guide, LayoutList, Meters, Unlock, Wallet } from '@open-condo/icons'
 import { useOrganization } from '@open-condo/next/organization'
+import { Button, Modal, Space, Typography } from '@open-condo/ui'
+import { colors } from '@open-condo/ui/dist/colors'
 
+import { FocusContainer } from '@condo/domains/common/components/FocusContainer'
 import { TourStep } from '@condo/domains/onboarding/utils/clientSchema'
 import { EXTERNAL_GUIDE_LINK } from '@condo/domains/onboarding/utils/clientSchema/constants'
 
@@ -36,6 +44,50 @@ type CompletedFlowModalDataType = {
     [key in TourStepTypeType]?: CompletedFlowModalDataValueType
 }
 
+const UnlockedFeature = ({ label }) => (
+    <Typography.Text size='medium' type='secondary'>
+        <Space size={4} direction='horizontal'>
+            <Unlock size='small' color={colors.green[5]}/>
+            {label}
+        </Space>
+    </Typography.Text>
+)
+
+const NewAbilitiesWrapper = styled.div`
+  padding: 24px;
+  border-radius: 12px;
+  background: ${colors.brandGradient[1]};
+  width: 100%;
+`
+const SuccessModal = styled(Modal)`
+  &.condo-modal > .condo-modal-content {
+    .condo-modal-header {
+      padding-right: 60px;
+    }
+    
+    & > .condo-modal-body {
+      padding-bottom: 20px;
+    }
+
+    & > .condo-modal-footer {
+      border-top: none;
+    }
+  }
+`
+const StyledFocusContainer = styled(FocusContainer)`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  margin: 0;
+  padding: 32px 24px;
+  text-align: center;
+
+  & img {
+    height: 240px;
+  }
+`
+
 type ComputedCompletedStepModalDataType = {
     title: string
     subtitleText: string
@@ -53,7 +105,7 @@ type ComputedCompletedFlowModalDataType = {
     title: string
 }
 
-export const useCompletedTourStepData = ({ activeStep, refetchSteps }) => {
+export const useCompletedTourModals = ({ activeStep, setActiveTourStep, refetchSteps }) => {
     const router = useRouter()
     const { organization } = useOrganization()
     const organizationId = useMemo(() => get(organization, 'id'), [organization])
@@ -104,7 +156,7 @@ export const useCompletedTourStepData = ({ activeStep, refetchSteps }) => {
         }
 
         await updateTourStep({ status: TourStepStatusType.Completed }, tourStep)
-    }, [activeStep, organizationId, refetchSteps, updateTourStep])
+    }, [activeStep, organizationId, refetchSteps, updateCompletedFlowModalData, updateTourStep])
 
 
     const completedStepModalDataDescription: CompletedStepModalDataType = useMemo(() => ({
@@ -293,10 +345,112 @@ export const useCompletedTourStepData = ({ activeStep, refetchSteps }) => {
     }), [activeStep, completedStepModalData])
 
 
-    return {
-        completedStepModalData: computedCompletedStepModalData,
+    return useMemo(() => ({
         updateCompletedStepModalData,
-        completedFlowModalData: computedCompletedFlowModalData,
-        updateCompletedFlowModalData,
-    }
+        CompletedStepModal: (
+            <Modal
+                open={!isEmpty(pickBy(computedCompletedStepModalData, Boolean))}
+                onCancel={() => {
+                    updateCompletedStepModalData(null)
+                    setActiveTourStep(null)
+                }}
+                width='big'
+                title={
+                    <Space size={8} direction='vertical'>
+                        <Typography.Title level={2}>{get(computedCompletedStepModalData, 'title')}</Typography.Title>
+                        <Space size={4} direction='horizontal'>
+                            <Typography.Text size='medium' type='secondary'>
+                                {get(computedCompletedStepModalData, 'subtitleText')}
+                            </Typography.Text>
+                            <Link href={get(computedCompletedStepModalData, 'subtitleLink', '')}>
+                                <Typography.Link size='medium'>
+                                    <div style={{ display: 'flex', gap: '8px', alignItems: 'center' }}>
+                                        {get(computedCompletedStepModalData, 'subtitleLinkLabel')}
+                                        {get(computedCompletedStepModalData, 'subtitleLinkIcon')}
+                                    </div>
+                                </Typography.Link>
+                            </Link>
+                        </Space>
+                    </Space>
+                }
+                footer={[
+                    <Button key='create' type='primary' onClick={() => {
+                        updateCompletedStepModalData(null)
+
+                        if (computedCompletedStepModalData.buttonOnClick) {
+                            computedCompletedStepModalData.buttonOnClick()
+                        }
+                    }}>
+                        {get(computedCompletedStepModalData, 'buttonLabel')}
+                    </Button>,
+                ]}
+            >
+                <Space size={40} direction='vertical'>
+                    <NewAbilitiesWrapper>
+                        <Row>
+                            {
+                                !isEmpty(computedCompletedStepModalData.newEmployeeFeatures) && (
+                                    <Col span={12}>
+                                        <Space size={20} direction='vertical'>
+                                            <Typography.Title level={4}>Теперь вы можете:</Typography.Title>
+                                            <Space size={16} direction='vertical'>
+                                                {
+                                                    computedCompletedStepModalData.newEmployeeFeatures.map(label => (
+                                                        <UnlockedFeature key={label} label={label} />
+                                                    ))
+                                                }
+                                            </Space>
+                                        </Space>
+                                    </Col>
+                                )
+                            }
+                            {
+                                !isEmpty(computedCompletedStepModalData.newResidentFeatures) && (
+                                    <Col span={12}>
+                                        <Space size={20} direction='vertical'>
+                                            <Typography.Title level={4}>Жители могут:</Typography.Title>
+                                            <Space size={16} direction='vertical'>
+                                                {
+                                                    computedCompletedStepModalData.newResidentFeatures.map(label => (
+                                                        <UnlockedFeature key={label} label={label} />
+                                                    ))
+                                                }
+                                            </Space>
+                                        </Space>
+                                    </Col>
+                                )
+                            }
+                        </Row>
+                    </NewAbilitiesWrapper>
+                    <Typography.Text>
+                        {get(computedCompletedStepModalData, 'bodyText')}
+                    </Typography.Text>
+                </Space>
+            </Modal>
+        ),
+        CompletedFlowModal: (
+            <SuccessModal
+                open={activeStep && !isEmpty(pickBy(computedCompletedFlowModalData, Boolean))}
+                title={get(computedCompletedFlowModalData, 'title')}
+                onCancel={() => updateCompletedFlowModalData(null)}
+                footer={[
+                    <Button key='tourButton' type='primary' onClick={async () => {
+                        updateCompletedFlowModalData(null)
+                        setActiveTourStep(null)
+
+                        await router.push('/tour')
+                    }}>
+                        Вернуться в путеводитель
+                    </Button>,
+                ]}
+            >
+                <StyledFocusContainer>
+                    <img src='/successDino.webp' width={240} height={240} />
+                </StyledFocusContainer>
+            </SuccessModal>
+        ),
+    }), [
+        activeStep, computedCompletedFlowModalData, computedCompletedStepModalData,
+        router, setActiveTourStep, updateCompletedFlowModalData, updateCompletedStepModalData,
+    ])
 }
